WIP: better docs for unified read/write functions [skip-ci]
